pyproject.toml
[project]
name = "mcp-oci-opsi"
version = "3.0.0"
description = "MCP server for Oracle Cloud Infrastructure Operations Insights - Enhanced with MCP and OCA Auth Bridge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "oci>=2.100.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
database = [
    "oracledb>=2.0.0",
]
auth = [
    "cryptography>=41.0.0",
    "PyJWT>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["."]
